+++ crates/module-system/module-implementations/sov-blob-storage/tests/integration/unregistered_sequencer.rs
@@ -2,7 +2,7 @@ use std::num::NonZero;
 
 use sov_blob_storage::{config_deferred_slots_count, config_unregistered_blobs_per_slot};
 use sov_mock_da::{MockAddress, MockBlob};
-use sov_modules_api::{Amount, CryptoSpec, SelectedBlob, Spec};
+use sov_modules_api::{Amount, CryptoSpec, Spec};
 use sov_modules_stf_blueprint::{BatchReceipt, Runtime};
 use sov_rollup_interface::da::RelevantBlobs;
 use sov_sequencer_registry::SequencerRegistry;
@@ -19,7 +19,7 @@ use crate::helpers_soft_confirmations::{
 use crate::{HashMap, TestData, S};
 
 fn make_unregistered_blobs<
-    RT: Runtime<S, BlobType = SelectedBlob<S>> + MinimalGenesis<S> + EncodeCall<SequencerRegistry<S>>,
+    RT: Runtime<S> + MinimalGenesis<S> + EncodeCall<SequencerRegistry<S>>,
 >(
     num_blobs: u64,
     sender: &TestSequencer<S>,
@@ -50,7 +50,7 @@ fn make_unregistered_blobs<
 }
 
 fn make_unregistered_blob_with_approx_size<
-    RT: Runtime<S, BlobType = SelectedBlob<S>> + MinimalGenesis<S> + EncodeCall<ValueSetter<S>>,
+    RT: Runtime<S> + MinimalGenesis<S> + EncodeCall<ValueSetter<S>>,
 >(
     sender: &TestSequencer<S>,
     size: usize,