+++ crates/module-system/sov-modules-stf-blueprint/src/sequencer_mode/unregistered.rs
@@ -2,20 +2,13 @@ use sov_modules_api::capabilities::{
     BatchFromUnregisteredSequencer, GasEnforcer, SequencerRemuneration, TransactionAuthorizer,
     UnregisteredAuthenticationError,
 };
-use sov_modules_api::{
-    Amount, BasicGasMeter, BatchSequencerOutcome, BatchSequencerReceipt, DaSpec, Gas, GasArray,
-    GasMeter, GasSpec, GetGasPrice, IgnoredTransactionReceipt, PreExecWorkingSet, Rewards, Spec,
-    StateProvider, WorkingSet,
-};
+use sov_modules_api::*;
 use tracing::{debug, warn};
 
 use crate::sequencer_mode::common::{
     apply_batch_logs, apply_tx, create_tx_receipt, get_gas_used, BatchReceipt,
 };
-use crate::{
-    ApplyTxResult, AuthTxOutput, IgnoredTxContents, Runtime, SkippedTxContents, StateCheckpoint,
-    TransactionReceipt, TxProcessingError, TxReceiptContents,
-};
+use crate::{ApplyTxResult, AuthTxOutput, Runtime, StateCheckpoint, TxReceiptContents};
 
 #[allow(clippy::result_large_err)]
 #[allow(clippy::too_many_arguments)]