+++ crates/module-system/sov-test-utils/src/rt_agnostic_blueprint.rs
@@ -10,9 +10,7 @@ use sov_modules_api::capabilities::{HasCapabilities, HasKernel};
 use sov_modules_api::execution_mode::Native;
 use sov_modules_api::prelude::axum::async_trait;
 use sov_modules_api::rest::{HasRestApi, StateUpdateReceiver};
-use sov_modules_api::{
-    CryptoSpec, NodeEndpoints, SelectedBlob, Spec, SyncStatus, ZkVerifier, Zkvm,
-};
+use sov_modules_api::{CryptoSpec, NodeEndpoints, Spec, SyncStatus, ZkVerifier, Zkvm};
 use sov_modules_rollup_blueprint::pluggable_traits::PluggableSpec;
 use sov_modules_rollup_blueprint::proof_serializer::SovApiProofSerializer;
 use sov_modules_rollup_blueprint::{FullNodeBlueprint, RollupBlueprint, SequencerCreationReceipt};
@@ -33,10 +31,7 @@ pub struct RtAgnosticBlueprint<S: Spec, R: RuntimeTrait<S>> {
 impl<S, R> RollupBlueprint<Native> for RtAgnosticBlueprint<S, R>
 where
     S: Spec + PluggableSpec,
-    R: RuntimeTrait<S>
-        + HasKernel<S, BlobType = SelectedBlob<S>>
-        + HasCapabilities<S>
-        + HasKernel<S, BlobType = SelectedBlob<S>>,
+    R: RuntimeTrait<S> + HasKernel<S> + HasCapabilities<S> + HasKernel<S>,
 {
     type Spec = S;
     type Runtime = R;
@@ -52,11 +47,7 @@ where
                 DefaultStorageSpec<<<S as Spec>::CryptoSpec as CryptoSpec>::Hasher>,
             >,
         > + PluggableSpec,
-    R: RuntimeTrait<S>
-        + HasRestApi<S>
-        + HasCapabilities<S>
-        + HasKernel<S, BlobType = SelectedBlob<S>>
-        + 'static,
+    R: RuntimeTrait<S> + HasRestApi<S> + HasCapabilities<S> + HasKernel<S> + 'static,
 {
     type DaService = StorableMockDaService;