+++ examples/demo-rollup/stf/src/runtime.rs
@@ -39,7 +39,7 @@ use sov_modules_api::capabilities::{Guard, HasCapabilities, HasKernel};
 #[cfg(feature = "native")]
 use sov_modules_api::macros::{expose_rpc, CliWallet};
 use sov_modules_api::prelude::*;
-use sov_modules_api::{DispatchCall, Event, Genesis, Hooks, MessageCodec, SelectedBlob, Spec};
+use sov_modules_api::{DispatchCall, Event, Genesis, Hooks, MessageCodec, Spec};
 
 #[cfg(feature = "native")]
 use crate::genesis_config::GenesisPaths;
@@ -154,7 +154,6 @@ impl<S: Spec> HasKernel<S> for Runtime<S>
 where
     S::Address: FromVmAddress<EthereumAddress>,
 {
-    type BlobType = SelectedBlob<S>;
     type Kernel<'a> = SoftConfirmationsKernel<'a, S>;
 
     fn inner(&mut self) -> Guard<Self::Kernel<'_>> {